Ensure to show the initial-hint the first time the pannable becomes scrollable
* src/hildon-pannable-area.c
(hildon_pannable_area_initial_effect): set to FALSE initial_effect
if the fadeout effect was used
(hildon_pannable_area_expose_event): check if it is required to
manage the initial effect on the begining of the function. Previously
it was made at the end, but this doesn't ensure that the scrollbars
will be painted on the first show. Do not set to FALSE initial_effect,
to ensure that the initial hint is showed the first time the pannable
becomes scrollable
* examples/Makefile.am
* examples/hildon-pannable-area-initial-hint-example.c:
New example, to show that the initial-hint is showed when the widget
becomes scrollable

Fixes: NB#116607 (Pannable area should always have scrollbar visible
initially if the application can be scrolled)