DateSelector documentation fixes
* hildon/hildon-date-selector.c
(hildon_date_selector_select_current_date)
(hildon_date_selector_get_date): Update documentation regarding days,
as the day of the month is dependent on both the month and the year.
(hildon_date_selector_get_date): Fix documentation of output arguments.

Fixes: MB#4698 (DateSelector documentation has bad output argument
description, and the day depends on both the month and the year)

Signed-off-by: Claudio Saavedra <csaavedra@igalia.com>