Don't emit HildonTouchSelector:changed unless necessary
When a HildonPickerDialog is cancelled, the original selection in its
HildonTouchSelector doesn't need to be restored in all cases.

If the dialog doesn't have a "Done" button then there's no way to
change its selection without closing it, thus it's not necessary to
restore anything.

This was causing an unnecessary emission of
HildonTouchSelector:changed than could confuse the application.

Fixes: NB#145156 (HildonPickerDialog: Entered Skype contact name
disappears from dialpad input box when call type dialog is opened and
cancelled)

Signed-off-by: Claudio Saavedra <csaavedra@igalia.com>