Make touchselector's live search filter by multiple strings matching
This makes the HildonLiveSearch in HildonTouchSelector match each
substring, separated by a space, in the entry, against the elements in
the model. This way, it is capable to match an entered "j hac" with
"Hacker, James".

Also, make use of a HildonLiveSearch::refilter signal handler to parse
the entered string, strip it, normalize it, and split it into the matching
tokens only once, and use these on the visibility method.